Updated Image and restructured App.jsx
